+++ packages/stacked/example/lib/app/app.router.dart
@@ -6,6 +6,7 @@
 
 // ignore_for_file: public_member_api_docs, unused_import, non_constant_identifier_names
 
+import 'package:example/app/custom_route_transition.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
@@ -51,8 +52,14 @@ class StackedRouter extends RouterBase {
   Map<Type, StackedRouteFactory> get pagesMap => _pagesMap;
   final _pagesMap = <Type, StackedRouteFactory>{
     HomeView: (data) {
+      var args = data.getArgs<HomeViewArguments>(
+        orElse: () => HomeViewArguments(),
+      );
       return MaterialPageRoute<dynamic>(
-        builder: (context) => const HomeView(),
+        builder: (context) => HomeView(
+          key: args.key,
+          title: args.title,
+        ),
         settings: data,
       );
     },
@@ -92,7 +99,7 @@ class StackedRouter extends RouterBase {
         pageBuilder: (context, animation, secondaryAnimation) =>
             const NonReactiveView(),
         settings: data,
-        transitionsBuilder: data.transition ?? TransitionsBuilders.slideBottom,
+        transitionsBuilder: data.transition ?? CustomRouteTransition.sharedAxis,
       );
     },
   };
@@ -101,6 +108,13 @@ class StackedRouter extends RouterBase {
 /// ************************************************************************
 /// Arguments holder classes
 /// *************************************************************************
+
+/// HomeView arguments holder class
+class HomeViewArguments {
+  final Key? key;
+  final String? title;
+  HomeViewArguments({this.key, this.title});
+}
 
 /// DetailsView arguments holder class
 class DetailsViewArguments {
@@ -121,6 +135,8 @@ class ExampleFormViewArguments {
 
 extension NavigatorStateExtension on NavigationService {
   Future<dynamic> navigateToHomeView({
+    Key? key,
+    String? title,
     int? id,
     bool preventDuplicates = true,
     Map<String, String>? parameters,
@@ -129,6 +145,7 @@ extension NavigatorStateExtension on NavigationService {
   }) async {
     return navigateTo(
       Routes.homeView,
+      arguments: HomeViewArguments(key: key, title: title),
       id: id,
       preventDuplicates: preventDuplicates,
       parameters: parameters,