+++ packages/stacked_generator/lib/src/generators/router/generator/route_generator_helper.dart
@@ -21,6 +21,7 @@ mixin RouteGeneratorHelper on BaseGenerator {
         validImports.where((element) => element.startsWith('package')).toSet();
     packageImports.add("package:stacked/stacked.dart");
     packageImports.add("package:stacked_services/stacked_services.dart");
+    if (routes.isNotEmpty) packageImports.add('package:flutter/material.dart');
     sortAndGenerate(packageImports);
     newLine();
 
@@ -292,10 +293,14 @@ mixin RouteGeneratorHelper on BaseGenerator {
     if (params.isNotEmpty) {
       params.forEach((param) {
         if (param.isRequired || param.isPositional) {
-          writeLine('required ${param.type} ${param.name},');
+          writeLine('required ${param.type} ${param.name}');
         } else {
-          writeLine('${param.type} ${param.name},');
+          writeLine('${param.type} ${param.name}');
         }
+        if (param.defaultValueCode != null) {
+          write(' = ${param.defaultValueCode}');
+        }
+        write(',');
       });
     }
     writeLine('''