+++ packages/stacked_generator/test/helpers/router_constants.dart
@@ -727,8 +727,8 @@ extension NavigatorStateExtension on NavigationService {
 
 Future<dynamic>
         navigateToTestClass( {
-null null,
-      int? id,
+null null
+,      int? id,
   bool preventDuplicates = true,
   Map<String, String>? parameters,
   Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)? transition,
@@ -1326,8 +1326,8 @@ extension NavigatorStateExtension on NavigationService {
 
 Future<dynamic>
         navigateToTestClass( {
-newType null,
-      int? id,
+newType null
+,      int? id,
   bool preventDuplicates = true,
   Map<String, String>? parameters,
   Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)? transition,
@@ -2228,6 +2228,7 @@ const kRouterTypeCupertinoWithNamePathNameClassName = '''
 // ignore_for_file: public_member_api_docs, unused_import, non_constant_identifier_names
 
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
 import 'package:stacked_services/stacked_services.dart';
 
@@ -2293,6 +2294,7 @@ const kRouterTypeCupertinoWithCupertinoNavTitle = '''
 // ignore_for_file: public_member_api_docs, unused_import, non_constant_identifier_names
 
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
 import 'package:stacked_services/stacked_services.dart';
 
@@ -2357,6 +2359,7 @@ const kRouterTypeCupertinoWithParametersPathFalseQueryFalse = '''
 // ignore_for_file: public_member_api_docs, unused_import, non_constant_identifier_names
 
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
 import 'package:stacked_services/stacked_services.dart';
 
@@ -2413,8 +2416,8 @@ extension NavigatorStateExtension on NavigationService {
 
 Future<dynamic>
         navigateToTestClass( {
-null null,
-      int? id,
+null null
+,      int? id,
   bool preventDuplicates = true,
   Map<String, String>? parameters,
   Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)? transition,
@@ -2438,6 +2441,7 @@ arguments: TestClassArguments(
 const kRouterTypeAdaptive = '''
 // ignore_for_file: public_member_api_docs, unused_import, non_constant_identifier_names
 
+import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
 import 'package:stacked_services/stacked_services.dart';
 
@@ -2501,6 +2505,7 @@ Future<dynamic>
 const kRouterTypeAdaptiveWithCupertinoNavTitle = '''
 // ignore_for_file: public_member_api_docs, unused_import, non_constant_identifier_names
 
+import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
 import 'package:stacked_services/stacked_services.dart';
 
@@ -2564,6 +2569,7 @@ Future<dynamic>
 const kRouterTypeAdaptiveWithParametersPathFalseQueryFalse = '''
 // ignore_for_file: public_member_api_docs, unused_import, non_constant_identifier_names
 
+import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
 import 'package:stacked_services/stacked_services.dart';
 
@@ -2620,8 +2626,8 @@ extension NavigatorStateExtension on NavigationService {
 
 Future<dynamic>
         navigateToTestClass( {
-null null,
-      int? id,
+null null
+,      int? id,
   bool preventDuplicates = true,
   Map<String, String>? parameters,
   Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)? transition,
@@ -3043,8 +3049,8 @@ extension NavigatorStateExtension on NavigationService {
 
 Future<dynamic>
         navigateToTestClass( {
-null null,
-      int? id,
+null null
+,      int? id,
   bool preventDuplicates = true,
   Map<String, String>? parameters,
   Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)? transition,
@@ -3275,8 +3281,8 @@ Future<dynamic>
 
 Future<dynamic>
         navigateToTestClass3( {
-null null,
-      int? id,
+null null
+,      int? id,
   bool preventDuplicates = true,
   Map<String, String>? parameters,
   Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)? transition,