+++ packages/stacked_generator/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.7.3
+- Fixes Import for Adaptive Route
+## 0.7.2
+- Add default values to Strong Type Navigation parameters
 ## 0.7.1
 - Add Strong Type Navigation parameters
 ## 0.7.0